@@ -57,40 +56,68 @@
     "estree-walker": "^3.0.3",
     "magic-string": "^0.30.8",
     "postcss": "^8.4.35",
-    "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.3",
+    "prettier": "^3.3.2",
+    "prettier-plugin-svelte": "^3.2.5",
     "sass": "^1.72.0",
-    "storybook": "^8.0.8",
-    "svelte": "5.0.0-next.138",
-    "svelte-check": "^3.6.9",
+    "storybook": "^8.1.10",
+    "svelte": "5.0.0-next.166",
+    "svelte-check": "^3.8.2",
     "svelte-preprocess-cssmodules": "github:DmitriVanGuard/svelte-preprocess-cssmodules#3b24f22",
     "tailwindcss": "^3.4.1",
-    "tslib": "^2.6.2",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.11",
+    "tslib": "^2.6.3",
+    "typescript": "^5.5.2",
+    "vite": "^5.3.1",
     "vitest": "^1.3.1"
   },
   "dependencies": {
     "@popsql/monaco-sql-languages": "^0.1.0",
     "@sveltejs/adapter-node": "^5.0.1",
+    "@tiptap/core": "^2.4.0",
+    "@tiptap/extension-bubble-menu": "^2.4.0",
+    "@tiptap/extension-bullet-list": "^2.4.0",
+    "@tiptap/extension-color": "^2.4.0",
+    "@tiptap/extension-heading": "^2.4.0",
+    "@tiptap/extension-highlight": "^2.4.0",
+    "@tiptap/extension-link": "^2.4.0",
+    "@tiptap/extension-list-item": "^2.4.0",
+    "@tiptap/extension-ordered-list": "^2.4.0",
+    "@tiptap/extension-paragraph": "^2.4.0",
+    "@tiptap/extension-placeholder": "^2.4.0",
+    "@tiptap/extension-text-style": "^2.4.0",
+    "@tiptap/extension-underline": "^2.4.0",
+    "@tiptap/html": "^2.4.0",
+    "@tiptap/pm": "^2.4.0",
+    "@tiptap/starter-kit": "^2.4.0",
+    "@tiptap/suggestion": "^2.4.0",
     "@xyflow/svelte": "^0.0.41",
     "chart.js": "^4.4.2",
     "esm-env": "^1.0.0",
+    "lucide-svelte": "^0.395.0",
     "monaco-editor": "^0.47.0",
-    "prettier-plugin-tailwindcss": "^0.5.11",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "prosemirror-state": "^1.4.3",
     "rxjs": "^7.8.1",
     "rxjs-exhaustmap-with-trailing": "^2.1.1",
     "san-studio": "github:santiment/san-studio#93d6d71",
     "san-webkit": "github:santiment/san-webkit#15afa8a",
+    "san-webkit-next": "github:santiment/san-webkit#lib-385ad266-260624",
     "svelte-runes": "^0.0.3",
     "tailwind-merge": "^2.2.1",
+    "tippy.js": "^6.3.7",
+    "tiptap-block-layout": "^0.0.6",
+    "tiptap-svelte-adapter": "^0.0.3",
     "ua-parser-js": "^1.0.37",
     "uuid": "^9.0.1",
+    "zeed-dom": "^0.13.3",
     "zod": "^3.22.4"
   },
   "pnpm": {
+    "overrides": {
+      "zeed-dom": "^0.13.3"
+    },
     "patchedDependencies": {
-      "@sveltejs/kit@2.5.10": "patches/@sveltejs__kit@2.5.10.patch"
+      "@sveltejs/kit@2.5.17": "patches/@sveltejs__kit@2.5.17.patch",
+      "san-webkit@0.0.1": "patches/san-webkit@0.0.1.patch"
     }
   }
 }
