@@ -2,14 +2,14 @@
   import type { OverridesDiff } from './utils'
   import Checkbox from '$lib/ui/Checkbox'
   import Parameter from '$lib/ui/Parameter'
-  import Svg from '$lib/ui/Svg.svelte'
+  import Svg from 'san-webkit-next/ui/core/Svg'
 
   let {
     widget,
     overridesDiff,
   }: { widget: App.Dashboard.QueryWidget; overridesDiff: OverridesDiff } = $props()
 
-  let queryParameters = $derived(Object.keys(widget.query.sqlQueryParameters))
+  let queryParameters = $derived(Object.keys(widget.data.sqlQueryParameters))
 
   function onToggle(parameter: string, isActive: boolean) {
     if (isActive) {
@@ -24,7 +24,7 @@
   <h4 class="flex items-center gap-2 fill-waterloo text-xs text-waterloo">
     <Svg id="query" w="12" />
 
-    {widget.query.name}
+    {widget.data.name}
   </h4>
 
   <parameters class="flex flex-wrap gap-6 gap-y-2">
@@ -35,7 +35,7 @@
         <Checkbox value={isActive} onclick={() => onToggle(parameter, isActive)} />
 
         <Parameter
-          parameter={{ key: parameter, value: widget.query.sqlQueryParameters[parameter] }}
+          parameter={{ key: parameter, value: widget.data.sqlQueryParameters[parameter] }}
         />
       </query-parameter>
     {/each}
