@@ -6,7 +6,6 @@
   import { useDashboardEditorCtx } from '$lib/Dashboard/ctx'
   import { useDataFlowSqlDataCtx } from '$lib/DataFlow/ctx/sqlData.svelte'
   import { showQueryWidgetFullscreenDialog$ } from './FullscreenDialog.svelte'
-  import AlertButton from './AlertButton.svelte'
 
   let {
     widget,
@@ -17,6 +16,7 @@
     onRefreshClick,
     onQueryChangesClick,
     sqlData,
+    onDeleteClick,
   }: {
     widget: App.Dashboard.QueryWidget
     id: number | string
@@ -26,19 +26,16 @@
     readonly?: boolean
     onRefreshClick: () => void
     onQueryChangesClick: () => void
+    onDeleteClick: () => void
   } = $props()
 
   const { dashboardEditor } = useDashboardEditorCtx()
   const { deleteDashboardQuery } = useDeleteDashboardQueryFlow()
-  const { changedParameters, mountRefreshPrompt } = useDataFlowSqlDataCtx()
+  // const { changedParameters, mountRefreshPrompt } = useDataFlowSqlDataCtx()
+  const changedParameters = new Set()
+  const mountRefreshPrompt = () => {}
+
   const showQueryWidgetFullscreenDialog = showQueryWidgetFullscreenDialog$()
-
-  function onDeleteClick() {
-    const dashboardId = dashboardEditor.id
-    if (!dashboardId) return
-
-    deleteDashboardQuery({ dashboardId, widget })
-  }
 
   function onFullscreenClick() {
     showQueryWidgetFullscreenDialog({ widget, sqlData })
@@ -53,8 +50,6 @@
   <h2 class="single-line mr-auto min-w-0 text-base">
     <a href="/query/{getSEOLinkFromIdAndTitle(id, name)}">{name}</a>
   </h2>
-
-  <AlertButton {widget}></AlertButton>
 
   {#if changedParameters.size > 0}
     <Button
