@@ -1,10 +1,23 @@
 import { redirect } from '@sveltejs/kit'
 import { getIdFromSEOLink } from 'san-webkit/lib/utils/url'
-import { UniQuery } from '$lib/api/index'
+import { BootFlag } from 'san-webkit-next/utils'
+import { UniQuery } from 'san-webkit-next/api/executor.js'
+import { queryGetCachedDashboardQueriesExecutions } from '$lib/Dashboard/flow/sqlData/api'
 import { queryGetDashboard } from './api'
 import { gotoDashboardPage } from './utils'
 
 export const ssr = false
+
+async function queryServerDashboardCache(
+  executor: ReturnType<typeof UniQuery>,
+  dashboardId: number,
+) {
+  if (BootFlag.get()) {
+    return null
+  }
+
+  return await queryGetCachedDashboardQueriesExecutions(executor)(dashboardId).catch(() => null)
+}
 
 export const load = async (event) => {
   const { slug = '' } = event.params
@@ -15,11 +28,17 @@ export const load = async (event) => {
     throw redirect(302, '/dashboard/edit/new')
   }
 
-  const preloaded = gotoDashboardPage.get()
-  const apiDashboard =
-    preloaded?.apiDashboard === undefined
-      ? await queryGetDashboard(UniQuery(event.fetch))(dashboardId).catch(() => null)
-      : preloaded.apiDashboard
+  const executor = UniQuery(event.fetch)
+
+  // const preloaded = gotoDashboardPage.get()
+  const [apiDashboard, serverDashboardCache] = await Promise.all([
+    queryGetDashboard(executor)(dashboardId).catch((e) => {
+      console.error(e)
+      return null
+    }),
+
+    queryServerDashboardCache(executor, dashboardId),
+  ])
 
   if (!apiDashboard) {
     throw redirect(302, '/dashboard/edit/new')
@@ -27,5 +46,6 @@ export const load = async (event) => {
 
   return {
     apiDashboard,
+    serverDashboardCache,
   }
 }
