@@ -1,6 +1,10 @@
 import { sveltekit } from '@sveltejs/kit/vite'
 import { defineConfig } from 'vitest/config'
 import { execSync } from 'node:child_process'
+import { createRequire } from 'module'
+import { WebkitSvg } from 'san-webkit-next/plugins/vite.js'
+
+globalThis.require = createRequire(import.meta.url)
 
 const mode = process.env.NODE_ENV
 const dev = mode !== 'production'
@@ -15,7 +19,7 @@ const GIT_HEAD =
   process.env.GIT_HEAD || execSync('git rev-parse HEAD').toString().trim().slice(0, 7)
 
 export default defineConfig({
-  plugins: [sveltekit()],
+  plugins: [sveltekit(), WebkitSvg()],
   test: {
     include: ['src/**/*.{test,spec}.{js,ts}'],
   },
